requirements.txt
matplotlib==3.9.2
numpy==1.24.3
Pillow==11.0.0
scikit_learn==1.5.2
streamlit==1.39.0
torch==2.5.0+cu124
torchvision==0.20.0+cu124
tqdm==4.66.5
